Updates after code review (part1)
